fix: change data type

+++ src/main/java/org/dainn/dainninventory/dto/OrderDetailDTO.java
@@ -3,19 +3,21 @@ package org.dainn.dainninventory.dto;
 import jakarta.validation.constraints.NotNull;
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
 public class OrderDetailDTO extends AbstractDTO{
     @NotNull(message = "Price is required")
-    private double price;
+    private BigDecimal price;
 
     @NotNull(message = "Quantity is required")
     private Integer quantity;
 
     @NotNull(message = "Total is required")
-    private double total;
+    private BigDecimal total;
 
     @NotNull(message = "Product is required")
     private Integer productId;
